+++ src/app/dropdown/dropdown-mock-with-long-data.ts
@@ -1,0 +1,18 @@
+export const MOCK_STATES = [
+  { value: 'AK', label: 'Alaska' },
+  { value: 'AZ', label: 'Arizona' },
+  { value: 'CA', label: 'California' },
+  { value: 'CO', label: 'Colorado' },
+  { value: 'MN', label: 'Minnesotasssssssssssssssssss Minnesotasssssssssssssssssss Minnesotasssssssssssssssssss' },
+  { value: 'ND', label: 'North Dakota' },
+  { value: 'OR', label: 'Oregon' },
+  { value: 'WA', label: 'Washington' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+  { value: 'WY', label: 'Wyoming' },
+];